PAR-2624: use the onInput msg for onClear in text inputs (#324)
Eliminates the `if-else` in the `view`. e.g.

```
if Maybe.isJust config.onClearInput then
...
else if config.hasSearchIcon then
...
else
...

```
I think one could not have both cross and search icon at the same time
previously due to that. Fixes to one single view.

Uses the webkit property for cross icon. This way, we can use the same
onInput msg on clear. not available in firefox.

Screenshots from chrome

![image](https://github.com/SGFinansAS/elm-components/assets/34012758/3afcb876-857a-4eb4-8a50-76a96b759336)

![image](https://github.com/SGFinansAS/elm-components/assets/34012758/a1c125d8-1fce-49e8-9ac4-e2df876e18fc)

This one is a bit problematic. Do we like the placement of the cross,
should we bring it left always or when there is currency or leave it
like this?

![image](https://github.com/SGFinansAS/elm-components/assets/34012758/0ff08801-4d63-43d8-bd15-6c96e3171cd8)