fix tooltip width and z-index issue (#353)
`min-width` was removed some time ago from tooltip and later it caused
issues for `DropdownFilter`, and as part of fix it was added again here
https://github.com/SGFinansAS/elm-components/pull/345

But the fix there also caused an issue in tooltip like below


![image](https://github.com/SGFinansAS/elm-components/assets/34012758/ea33b1ed-fa9c-4c10-8b69-b1067cbfb846)

The tooltip lost its position at the center, and `min-width` caused the
tooltip to cover a very large area (as big as the element that is being
tooltipped). This is bad because you cannot click anything because of
z-index.

Now with this fix, both dropdown and tooltip should work as intended.


![image](https://github.com/SGFinansAS/elm-components/assets/34012758/6070a5c8-0e13-488f-8918-7dbcfac70224)

![image](https://github.com/SGFinansAS/elm-components/assets/34012758/877d86ab-f476-4470-914d-c89d23a4ad4a)