Add helper functions for CSS properties related to grids (#363)
Add helper functions for properties that are related to CSS Grids.

Elm-css is a bit outdated and does not support newer CSS syntax. This is
a continuation of @morteako 's idea of adding the missing functions to
the component library (`gap` and `gap2` in his case).

I haven't done the effort of typing these functions in a sophisticated
way, so most just accept a single _string_ parameter. Adding better
types would be a greater effort since the Grid spec introduces a bunch
of new units that are also not present in elm-css. Anyway, I think it
adds some value to be able to type:

```elm
css
  [ displayGrid
  , gridTemplateColumns "repeat(4, 1fr)"
  ]
```

Instead of:

```elm
css
  [ Css.property "display" "grid"
  , Css.property "grid-template-columns" "repeat(4, 1fr)"
  ]
```

Added some links to MDN as well to get some help right in the editor.