Workaround for Chrome bug in details element (#429)
This is a workaround for a bug in Chrome that messes up the Grid layout
when using the HTML details element together with the `display:
contents` rule. The bug is currently apparent on the Supplier Invoice
and Change of Party pages in FinansFront.

Fixed as described on the Chrome bug tracking site:
https://issues.chromium.org/issues/40069963

Before:


![image](https://github.com/user-attachments/assets/fdc4c18b-98b9-46b3-b278-9bcdb00d5f48)

After:


![image](https://github.com/user-attachments/assets/1a020744-c382-4df6-be07-631ac42b2e6a)