Add groups to multiselect (#461)
<img width="1142" alt="image"
src="https://github.com/user-attachments/assets/849511ac-1dd4-4dbd-9426-1682f1da6ad5"
/>

Adds groups to multiselectdropdown. See design on left.

I have not changed the actual design and spacing of the select items, so
it is not 100% as the design.

I have kept the current withOptions-functions, so it should be
non-breaking