MultiSelectDropdown - add filter (#463)
Add filtering to MultiSelectDropdown.
This is just an extra option, so the current uses of MultiSelectDropdown
should not be affected


## With filter
![2025-04-25 10 41
38](https://github.com/user-attachments/assets/9aba2cbc-9f5c-4326-8044-8ed8a64a2d9e)

## "Old" - without filter
![2025-04-25 10 50
55](https://github.com/user-attachments/assets/2f9fa354-baf2-43c1-9c1c-0f1d94bfe8f3)

(the hint is bit weirdly placed, but that is a current behaviour that is
not related to this PR)

also add `attrEmpty` 

Also removes important from label boxshadow css 👀 
Will need to check if that is ok.
Another option is make it toggleable in the label i guess? 


I tried to a add a cross-icon to clear the filter, but it had some
focus?-issues so it didnt work, so i scoped it out of this PR: can be
added later